Moved proper Methos to BradcastDomain
The Hierachy set up and the clean topology
are suitable methods for broadcastdomain class
